requirements.txt
setuptools
matplotlib
sympy
tqdm
matplotlib
numpy
scikit-learn
scikit-optimize 
scipy
wandb
seaborn
deepxde
simpy
hydra-core
matlabengine